voice-agent/requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]>=0.12.0
supabase>=2.0.0
python-dotenv>=1.0.0
